2_openai/deep_research_personal/requirements.txt
gradio
openai
python-dotenv
sendgrid
openai-agents